+++ packages/yourloops/components/header-bars/patient-nav-bar.tsx
@@ -53,7 +53,7 @@ const PatientNavBar: FunctionComponent<PatientNavBarProps> = (props) => {
   } = props
 
   const theme = useTheme()
-  const isMobile = useMediaQuery(theme.breakpoints.down('sm'));
+  const isMobile = useMediaQuery(theme.breakpoints.down('sm'))
 
   return (
     <Box